requirements.txt
Django==2.0.2
psycopg2-binary==2.7.4
django-reversion==2.0.13
pyLibravatar
py3dns
gunicorn
flake8
